refactor so that halt_error and all the halt_{status name} methods are in a sharable module. rails just defines #halt, the Halt error class, and #handle_halt. move other rails-specific files to a rails folder.
