+++ lib/api_hammer/rails/check_required_params.rb
